create: Move LogicalRecordPositionIndex into index submod
